Some minor refactorings
